app/requirements.txt
Django==2.2.10
gunicorn==19.9.0
psycopg2==2.8.3
djangorestframework==3.10.3
markdown==3.1.1
django-filter==2.2.0
django-grappelli==2.13.2
django-countries==5.5
Pillow==6.2.2
django-imagekit==4.0.2
django-rest-swagger==2.2.0
django-countries==5.5
django-extensions==2.2.5
Werkzeug==0.16.0
django-admin-autocomplete-filter==0.4
django-import-export==2.0.1
lxml>=3.5.0
python-docx==0.8.10
babel==2.8.0